requirements.txt
pytest == 7.4.0
selenium == 4.10.0
faker == 19.3.0
requests
allure-pytest